requirements.txt
dnspython
python-whois
beautifulsoup4
requests